feat: add support for importing a solution
On realization that a challenge benefits from some things you used
in previous solutions, you may find it helpful to copy-and-paste
some of the source files from the relevant directories.

However, some CTFs are progressive, and many will have a flag that is
particularly related to a previous challenge. This offers a shortcut
to replace all the files in a solution directory with those of a
previous one (specifiable in the FROM variable which defaults to the
last one).

Signed-off-by: Sam Stuewe <samuel.stuewe@gmail.com>